requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
streamlit>=1.35.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
pydantic>=2.0.0

# LangChain
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.2.0
langchain-text-splitters>=0.3.0
langchain-community>=0.3.0
chromadb>=0.5.0

# Document parsing
markitdown>=0.0.1
pypdf>=4.0.0

# Evaluation
ragas>=0.2.0

soupsieve~=2.8.3
beautifulsoup4~=4.14.3
typing_extensions~=4.15.0
lxml~=6.0.2
GitPython~=3.1.50
pathlib~=1.0.1
gitdb~=4.0.12
configparser~=7.2.0
h11~=0.16.0
cryptography~=48.0.0
defusedxml~=0.7.1
numpy~=1.26.4
pillow~=10.4.0
olefile~=0.47
pip~=25.3
attrs~=26.1.0
distro~=1.9.0
protobuf~=6.33.6
filelock~=3.25.2
cffi~=2.0.0
setuptools~=81.0.0
dill~=0.4.1
grpcio~=1.80.0
msal~=1.36.0
python-pptx~=1.0.2
xlsxwriter~=3.2.9
rich~=13.9.4
Pygments~=2.20.0
markdown-it-py~=4.0.0
toml~=0.10.2
python-telegram-bot~=22.5
requests~=2.33.1
colorama~=0.4.6
PyYAML~=6.0.3
idna~=3.11
multidict~=6.7.1
propcache~=0.4.1
pydantic_core~=2.41.5
zipp~=3.23.1
sniffio~=1.3.1
azure-core~=1.41.0
isodate~=0.7.2
aiohttp~=3.13.5
urllib3~=2.5.0
msal-extensions~=1.3.1
importlib_metadata~=8.7.1
pyproject_hooks~=1.2.0
build~=1.5.0
smmap~=5.0.3
click~=8.3.1
httpcore~=1.0.9
httpx~=0.28.1
zstandard~=0.25.0
jiter~=0.12.0
mdurl~=0.1.2
charset-normalizer~=3.4.4
overrides~=7.7.0
scipy~=1.17.1
instructor~=1.15.1
langchain-core~=0.3.86
typer~=0.25.1
tqdm~=4.67.1
tiktoken~=0.12.0
datasets~=4.8.5
pandas~=2.3.3
langsmith~=0.8.3
appdirs~=1.4.4
tenacity~=8.5.0
regex~=2026.4.4
threadpoolctl~=3.6.0
torch~=2.11.0
packaging~=24.2
mpmath~=1.3.0
sympy~=1.14.0
limits~=5.6.0
shellingham~=1.5.4
altair~=5.5.0
jsonschema~=4.26.0
narwhals~=2.21.0
pyarrow~=24.0.0
Jinja2~=3.1.6
referencing~=0.37.0
fsspec~=2026.2.0
yarl~=1.23.0
MarkupSafe~=3.0.3
magika~=0.6.3
onnxruntime~=1.20.1
openai~=2.34.0
websockets~=16.0
anyio~=4.11.0
orjson~=3.11.8
xlrd~=2.0.2
openpyxl~=3.1.5
SQLAlchemy~=2.0.44
python-dateutil~=2.9.0.post0
pydeck~=0.9.2
PyPika~=0.51.1
xxhash~=3.7.0
aiosignal~=1.4.0
frozenlist~=1.8.0
aiohappyeyeballs~=2.6.1
certifi~=2025.11.12
starlette~=0.38.6
cobble~=0.1.4
mammoth~=1.12.0
tzdata~=2025.3
httptools~=0.7.1
watchfiles~=1.1.1
importlib_resources~=7.1.0
pybase64~=1.4.3
bcrypt~=5.0.0
kubernetes~=35.0.0
mmh3~=5.2.1
chroma-hnswlib~=0.7.6
pydantic-settings~=2.14.0
huggingface_hub~=0.36.2
multiprocess~=0.70.19
nibabel~=5.4.2
six~=1.17.0
greenlet~=3.3.0
networkx~=3.6.1
pydot~=4.0.1
oauthlib~=3.3.1
blinker~=1.9.0
et_xmlfile~=2.0.0
Deprecated~=1.3.1
annotated-types~=0.7.0
simplejson~=3.20.2
tornado~=6.5.5
watchdog~=6.0.0
langgraph~=1.1.10
ormsgpack~=1.12.2
uuid_utils~=0.14.1
requests-toolbelt~=1.0.0
puremagic~=2.1.0
itsdangerous~=2.2.0
cachetools~=5.5.2
websocket-client~=1.9.0
pyreadline3~=3.5.4
docstring_parser~=0.18.0
jsonpointer~=3.1.1
jsonschema-specifications~=2025.9.1
durationpy~=0.10
requests-oauthlib~=2.0.0
markdownify~=1.2.2
pydub~=0.25.1
SpeechRecognition~=3.16.1
youtube-transcript-api~=1.2.4
tokenizers~=0.22.2
Werkzeug~=3.1.4
marshmallow~=3.26.2
flatbuffers~=25.12.19
transformers~=4.57.6
jsonpatch~=1.33
safetensors~=0.7.0
typing-inspect~=0.9.0
dataclasses-json~=0.6.7
mypy_extensions~=1.1.0
rdflib~=7.6.0
httpx-sse~=0.4.3
sentence-transformers~=3.0.0
typing_extensions~=4.15.0
pydantic_core~=2.41.5
importlib_metadata~=8.7.1
pyproject_hooks~=1.2.0
importlib_resources~=7.1.0
huggingface_hub~=0.36.2
et_xmlfile~=2.0.0
uuid_utils~=0.14.1
docstring_parser~=0.18.0
mypy_extensions~=1.1.0